Implement logic that allows uploading and downloading empty files
These are zero length S3 objects with a name ending in a slash.

We record the metadata about ownership, timestamps about these folders

+++ command/cp.go
@@ -491,11 +491,11 @@ func (c Copy) Run(ctx context.Context) error {
 	}
 
 	for object := range objch {
-		if errorpkg.IsCancelation(object.Err) || object.Type.IsDir() {
+		if errorpkg.IsCancelation(object.Err) {
 			continue
 		}
 
-		if !object.Type.IsRegular() {
+		if !object.Type.IsRegular() && !object.Type.IsDir() {
 			err := fmt.Errorf("object '%v' is not a regular file", object)
 			merrorObjects = multierror.Append(merrorObjects, err)
 			printError(c.fullCommand, c.op, err)
@@ -541,7 +541,7 @@ func (c Copy) Run(ctx context.Context) error {
 		case srcurl.Type == c.dst.Type: // local->local or remote->remote
 			task = c.prepareCopyTask(ctx, srcurl, c.dst, isBatch, c.metadata)
 		case srcurl.IsRemote(): // remote->local
-			task = c.prepareDownloadTask(ctx, srcurl, c.dst, isBatch)
+			task = c.prepareDownloadTask(ctx, srcurl, c.dst, isBatch, object.Type.IsDir())
 		case c.dst.IsRemote(): // local->remote
 			task = c.prepareUploadTask(ctx, srcurl, c.dst, isBatch, c.metadata)
 		default:
@@ -583,9 +583,10 @@ func (c Copy) prepareDownloadTask(
 	srcurl *url.URL,
 	dsturl *url.URL,
 	isBatch bool,
+	srcIsDir bool,
 ) func() error {
 	return func() error {
-		dsturl, err := prepareLocalDestination(ctx, srcurl, dsturl, c.flatten, isBatch, c.storageOpts)
+		dsturl, err := prepareLocalDestination(ctx, srcurl, dsturl, c.flatten, isBatch, c.storageOpts, srcIsDir)
 		if err != nil {
 			return err
 		}
@@ -644,33 +645,51 @@ func (c Copy) doDownload(ctx context.Context, srcurl *url.URL, dsturl *url.URL) 
 		}
 		return err
 	}
+	// Check to see if the source is a directory for locally creation a directory too
+	srcObj, err := srcClient.Stat(ctx, srcurl)
+	if err != nil {
+		var objNotFound *storage.ErrGivenObjectNotFound
+		if !errors.As(err, &objNotFound) {
+			return err
+		}
+
+	}
 
 	dstPath := filepath.Dir(dsturl.Absolute())
 	dstFile := filepath.Base(dsturl.Absolute())
-	file, err := dstClient.CreateTemp(dstPath, dstFile)
-	if err != nil {
-		return err
-	}
 
-	writer := newCountingReaderWriter(file, c.progressbar)
-	size, err := srcClient.Get(ctx, srcurl, writer, c.concurrency, c.partSize)
-	file.Close()
-
-	if err != nil {
-		dErr := dstClient.Delete(ctx, &url.URL{Path: file.Name(), Type: dsturl.Type})
-		if dErr != nil {
-			printDebug(c.op, dErr, srcurl, dsturl)
+	isDir := srcObj.Type.IsDir()
+	var size int64 = 0
+	if isDir {
+		err = dstClient.CreateDir(ctx, dsturl.Absolute(), storage.Metadata{})
+		if err != nil {
+			return err
 		}
-		return err
+	} else {
+		file, err := dstClient.CreateTemp(dstPath, dstFile)
+		if err != nil {
+			return err
+		}
+
+		writer := newCountingReaderWriter(file, c.progressbar)
+		size, err = srcClient.Get(ctx, srcurl, writer, c.concurrency, c.partSize)
+
+		file.Close()
+		if err != nil {
+			dErr := dstClient.Delete(ctx, &url.URL{Path: file.Name(), Type: dsturl.Type})
+			if dErr != nil {
+				printDebug(c.op, dErr, srcurl, dsturl)
+			}
+			return err
+		}
+		err = dstClient.Rename(file, dsturl.Absolute())
+		if err != nil {
+			return err
+		}
 	}
 
 	if c.deleteSource {
 		_ = srcClient.Delete(ctx, srcurl)
-	}
-
-	err = dstClient.Rename(file, dsturl.Absolute())
-	if err != nil {
-		return err
 	}
 
 	if c.preserveOwnership {
@@ -786,7 +805,15 @@ func (c Copy) doUpload(ctx context.Context, srcurl *url.URL, dsturl *url.URL, ex
 	}
 
 	reader := newCountingReaderWriter(file, c.progressbar)
-	err = dstClient.Put(ctx, reader, dsturl, metadata, c.concurrency, c.partSize)
+	fi, err := file.Stat()
+	if err != nil {
+		return err
+	}
+	if fi.IsDir() {
+		err = dstClient.CreateDir(ctx, dsturl, metadata)
+	} else {
+		err = dstClient.Put(ctx, reader, dsturl, metadata, c.concurrency, c.partSize)
+	}
 
 	if err != nil {
 		return err
@@ -957,6 +984,10 @@ func prepareRemoteDestination(
 		objname = srcurl.Relative()
 	}
 
+	if objname == "." {
+		return dsturl
+	}
+
 	if dsturl.IsPrefix() || dsturl.IsBucket() {
 		dsturl = dsturl.Join(objname)
 	}
@@ -972,6 +1003,7 @@ func prepareLocalDestination(
 	flatten bool,
 	isBatch bool,
 	storageOpts storage.Options,
+	srcIsDir bool,
 ) (*url.URL, error) {
 	objname := srcurl.Base()
 	if isBatch && !flatten {
@@ -1008,11 +1040,11 @@ func prepareLocalDestination(
 		if err != nil {
 			return nil, err
 		}
-		if strings.HasSuffix(dsturl.Absolute(), "/") {
+		if strings.HasSuffix(dsturl.Absolute(), "/") && !srcIsDir {
 			dsturl = dsturl.Join(objname)
 		}
 	} else {
-		if obj.Type.IsDir() {
+		if obj.Type.IsDir() && !srcIsDir {
 			dsturl = obj.URL.Join(objname)
 		}
 	}
@@ -1058,7 +1090,7 @@ func validateCopyCommand(c *cli.Context) error {
 	}
 
 	// we don't operate on S3 prefixes for copy and delete operations.
-	if srcurl.IsBucket() || srcurl.IsPrefix() {
+	if srcurl.IsBucket() {
 		return fmt.Errorf("source argument must contain wildcard character")
 	}
 
@@ -1105,6 +1137,10 @@ func validateUpload(ctx context.Context, srcurl, dsturl *url.URL, storageOpts st
 	obj, err := srcclient.Stat(ctx, srcurl)
 	if err != nil {
 		return err
+	}
+
+	if obj.Type.IsDir() {
+		return nil
 	}
 
 	// 'cp dir/ s3://bucket/prefix-without-slash': expect a trailing slash to
